main_app/requirements.txt
aiosignal==1.2.0
alembic==1.7.4
anyio==3.3.4
asgiref==3.4.1
async-timeout==4.0.2
asyncpg==0.25.0
attrs==21.4.0
bcrypt==3.2.0
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.3
clikit==0.6.2
ecdsa==0.17.0
fastapi==0.73.0
frozenlist==1.3.0
h11==0.12.0
httptools==0.3.0
idna==3.3
Jinja2==3.0.3
Mako==1.1.5
MarkupSafe==2.0.1
multidict==6.0.2
passlib==1.7.4
pastel==0.2.1
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
pylev==1.4.0
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.0.1
uvicorn==0.17.4
uvloop==0.16.0
watchgod==0.7
websockets==10.1
yarl==1.7.2
